Solution-6
I added the title, drawer, and button. I didnt get to format everything because I spent a lot of time on actually getting the intial example to run. I still had to fix the toggle for the button and hamburger button cause they were linked and then style it a little better.