Fixed Solution Day 6
Created Title bar with button that opens drawer on the side. Linked the about button in the drawer to pull up the info window, put in a dummy button in the drawer that does nothing, and also put in a close button to close the drawer. Put in a Click Me! button that also displays the pop up info. (I could not figure out how to space the button from the navbar however, which I spent the majority of my time on...please help)